requirements.txt
Django==4.0.1
environs[django]==9.3.2
djangorestframework==3.13.1
django-tinymce==3.4.0
Pillow==9.2.0
django-allauth==0.51.0
djangorestframework==3.13.1
requests==2.28.1

